fix: move into a view directory and change export method
