+++ packages/runtime-react/src/shuvi-app/index.browser.ts
@@ -1,0 +1,7 @@
+declare const __BROWSER__: boolean;
+
+if (__BROWSER__) {
+  module.exports = require('./view/clientView.browser');
+} else {
+  module.exports = require('./view/serverView');
+}
