+++ packages/runtime-react/src/shuvi-app/index.hash.ts
@@ -1,0 +1,5 @@
+if (__BROWSER__) {
+  module.exports = require('./view/clientView.hash');
+} else {
+  module.exports = require('./view/serverView');
+}
