+++ packages/runtime-react/src/shuvi-app/view/ReactView.server.tsx
@@ -2,14 +2,14 @@ import React from 'react';
 import { renderToString } from 'react-dom/server';
 import { Runtime } from '@shuvi/types';
 import { Router } from 'react-router-dom';
-import { createServerHistory } from './router/history';
-import { setHistory } from './router/router';
-import Loadable, { LoadableContext } from './loadable';
-import AppContainer from './AppContainer';
-import { IReactServerView, IReactAppData } from './types';
-import { Head } from './head';
-import { createRedirector } from './utils/createRedirector';
-import { matchRoutes } from './router/matchRoutes';
+import { createServerHistory } from '../router/history';
+import { setHistory } from '../router/router';
+import Loadable, { LoadableContext } from '../loadable';
+import AppContainer from '../AppContainer';
+import { IReactServerView, IReactAppData } from '../types';
+import { Head } from '../head';
+import { createRedirector } from '../utils/createRedirector';
+import { matchRoutes } from '../router/matchRoutes';
 
 import IAppComponent = Runtime.IAppComponent;
 import IRouteComponent = Runtime.IRouteComponent;
