+++ packages/runtime-react/src/shuvi-app/view/clientView.browser.ts
@@ -1,0 +1,4 @@
+import { ReactClientView } from './ReactView.client';
+import { createBrowserHistory } from '../router/history';
+
+export default new ReactClientView(createBrowserHistory);
