+++ packages/runtime-react/src/shuvi-app/view/clientView.hash.ts
@@ -1,0 +1,4 @@
+import { ReactClientView } from './ReactView.client';
+import { createHashHistory } from '../router/history';
+
+export default new ReactClientView(createHashHistory);
