+++ packages/runtime-react/src/shuvi-app/view/serverView.ts
@@ -1,0 +1,3 @@
+import { ReactServerView } from './ReactView.server';
+
+export default new ReactServerView();
